+++ test/devices_tests/weather_test.py
@@ -43,6 +43,7 @@ class TestWeather(unittest.TestCase):
             group_address_brightness_south="1/3/6",
             group_address_brightness_west="1/3/7",
             group_address_brightness_north="1/3/8",
+            group_address_temperature="1/4/4",
         )
 
         weather._brightness_east.payload = DPTArray(
@@ -369,3 +370,9 @@ class TestWeather(unittest.TestCase):
         weather = Weather(name="weather", xknx=xknx, group_address_temperature="1/3/4")
         self.assertTrue(weather._temperature.has_group_address(GroupAddress("1/3/4")))
         self.assertFalse(weather._temperature.has_group_address(GroupAddress("1/2/4")))
+
+    def test_unique_id(self):
+        """Test unique id functionality."""
+        xknx = XKNX()
+        weather = Weather(name="weather", xknx=xknx, group_address_temperature="1/3/4")
+        self.assertEqual(weather.unique_id, "1/3/4")