+++ xknx/devices/cover.py
@@ -138,6 +138,11 @@ class Cover(Device):
         yield self.position_target
         yield self.angle
 
+    @property
+    def unique_id(self) -> Optional[str]:
+        """Return unique id for this device."""
+        return f"{self.updown.group_address}"
+
     @classmethod
     def from_config(cls, xknx: "XKNX", name: str, config: Dict[str, Any]) -> "Cover":
         """Initialize object from configuration structure."""