+++ xknx/devices/device.py
@@ -58,6 +58,11 @@ class Device(ABC):
         # yield from (<list all used RemoteValue instances>)
         yield from ()
 
+    @property
+    def unique_id(self) -> Optional[str]:
+        """Return local unique id of this device."""
+        return None
+
     def register_device_updated_cb(self, device_updated_cb: DeviceCallbackType) -> None:
         """Register device updated callback."""
         self.device_updated_cbs.append(device_updated_cb)