+++ changelog.md
@@ -2,7 +2,9 @@
 
 ## Unreleased changes
 
--
+### HA integration
+
+- Add experimental (opt-in) support for unique ids
 
 ## 0.17.4 Bugfix for ValueReader 2021-03-26